Flavorverse_Flask_API/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Mail==0.9.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
requests==2.29.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.1.3
WTForms==3.2.1
six==1.16.0
